Feature extraction update

+++ enigma/pretrains.py
@@ -31,14 +31,17 @@ def prepare(rkeys):
 
 def translate(f_cnf, f_conj, f_out):
    out = file(f_out, "w")
-   f_empty = "empty.tmp"
-   os.system("rm -fr %s" % f_empty)
-   os.system("touch %s" % f_empty)
-   subprocess.call(["enigma-features", "--free-numbers", f_cnf, f_empty, f_conj], \
-      stdout=out)
-      #stdout=out, stderr=subprocess.STDOUT)
+   if not f_conj:
+      subprocess.call(["enigma-features", "--free-numbers", f_cnf], stdout=out)
+   else:   
+      f_empty = "empty.tmp"
+      os.system("rm -fr %s" % f_empty)
+      os.system("touch %s" % f_empty)
+      subprocess.call(["enigma-features", "--free-numbers", f_cnf, f_empty, f_conj], \
+         stdout=out)
+         #stdout=out, stderr=subprocess.STDOUT)
+      os.system("rm -fr %s" % f_empty)
    out.close()
-   os.system("rm -fr %s" % f_empty)
 
 def make(rkeys, out=None):
    pre = []
